app/requirements.txt
alembic==1.7.3
amqp==5.0.6
aniso8601==9.0.1
apispec==4.5.0
attrs==21.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
flasgger==0.9.5
Flask==2.0.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
kombu==5.1.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.1
mistune==0.8.4
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
PyJWT==1.4.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.17
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
